Finishing the post search

+++ backend/delivery/controller/material_controller.go
@@ -1,6 +1,7 @@
 package controller
 
 import (
+	"fmt"
 	"net/http"
 	"strconv"
 	"time"
@@ -90,6 +91,7 @@ func (mc *MaterialController) GetMaterialByID(ctx *gin.Context) {
 
 func (mc *MaterialController) CreateMaterial(ctx *gin.Context) {
 	// Grab user ID
+
 	userID, exist := ctx.Get("user_id")
 	if !exist {
 		ctx.JSON(400, gin.H{"error": "User ID not found in context"})
@@ -107,11 +109,11 @@ func (mc *MaterialController) CreateMaterial(ctx *gin.Context) {
 		return
 	}
 
-	// Collect the files
-	ctx.Request.Body = http.MaxBytesReader(ctx.Writer, ctx.Request.Body, 20<<20) // 20 MB limit
+	ctx.Request.Body = http.MaxBytesReader(ctx.Writer, ctx.Request.Body, 20<<20) // 10 MB limit
 	form, err := ctx.MultipartForm()
 	if err != nil {
-		ctx.JSON(400, gin.H{
+		fmt.Println("Error parsing multipart form:", err)
+		ctx.JSON(http.StatusBadRequest, gin.H{
 			"error":   "Failed to parse multipart form",
 			"details": err.Error(),
 		})
@@ -119,38 +121,88 @@ func (mc *MaterialController) CreateMaterial(ctx *gin.Context) {
 	}
 
 	var material models.Materials
-	if err := ctx.ShouldBind(&material); err != nil {
-		ctx.JSON(400, gin.H{
-			"error":   "Invalid or missing form data. Ensure all required fields are provided and correctly formatted.",
-			"details": err.Error(),
-		})
+
+	departmentID := form.Value["department_id"]
+	if len(departmentID) == 0 {
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Department ID is required"})
 		return
 	}
+	depId, err := primitive.ObjectIDFromHex(departmentID[0])
+	if err != nil {
+		fmt.Println("Invalid Department ID format:", err)
+		ctx.JSON(400, gin.H{"error": "Invalid Department ID format"})
+		return
+	}
+	material.DepartmentID = depId
+
+	title := form.Value["title"]
+	if len(title) == 0 {
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Title is required"})
+		return
+	}
+	material.Title = title[0]
+
+	year := form.Value["year"]
+	if len(year) == 0 {
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Year is required"})
+		return
+	}
+	yearInt, err := strconv.Atoi(year[0])
+	if err != nil || yearInt < 1 {
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Year must be a positive integer"})
+		return
+	}
+	material.Year = yearInt
+
+	semester := form.Value["semester"]
+	if len(semester) == 0 {
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Semester is required"})
+		return
+	}
+	semesterInt, err := strconv.Atoi(semester[0])
+	if err != nil || (semesterInt != 1 && semesterInt != 2) {
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Semester must be either 1 or 2"})
+		return
+	}
+	material.Semester = semesterInt
 
 	material.UploadedBy = obId
 
+	if material.DepartmentID.IsZero() {
+		fmt.Println("Department ID is required")
+		ctx.JSON(400, gin.H{"error": "Department ID is required"})
+		return
+	}
+	if material.Year < 1 {
+		fmt.Println("Year must be a positive integer")
+		ctx.JSON(400, gin.H{"error": "Year must be a positive integer"})
+		return
+	}
+	if material.Semester < 1 || material.Semester > 2 {
+		fmt.Println("Semester must be either 1 or 2")
+		ctx.JSON(400, gin.H{"error": "Semester must be either 1 or 2"})
+		return
+	}
 	files := form.File["file"]
-	if len(files) == 0 {
-		ctx.JSON(400, gin.H{"error": "No files uploaded"})
+	if len(files) == 0 && len(material.FileURL) > 1 {
+		fmt.Println("You have to upload one file")
+		ctx.JSON(400, gin.H{"error": "You have to upload one file"})
 		return
 	}
 
 	for _, file := range files {
 		if file.Header.Get("Content-Type") != "application/pdf" {
+			fmt.Println("Only PDF files are allowed")
 			ctx.JSON(400, gin.H{"error": "Only PDF files are allowed"})
 			return
 		}
-	}
-
-	if len(files) > 1 {
-		ctx.JSON(400, gin.H{"error": "Only one file can be uploaded at a time"})
-		return
 	}
 
 	// TODO: Implement file storage logic
 	urls, err := mc.StorageUsecase.UploadFile(files)
 
 	if err != nil {
+		fmt.Println("Failed to upload files:", err)
 		ctx.JSON(500, gin.H{
 			"error":   "Failed to upload files",
 			"details": err.Error(),
@@ -162,6 +214,9 @@ func (mc *MaterialController) CreateMaterial(ctx *gin.Context) {
 
 	newMaterial, err := mc.MaterialUsecase.CreateMaterial(ctx, material)
 	if err != nil {
+		mc.StorageUsecase.DeleteFile(urls) // Clean up uploaded file if creation fails
+
+		fmt.Println("Failed to create material:", err)
 		ctx.JSON(500, gin.H{"error": "Failed to create material"})
 		return
 	}
@@ -213,6 +268,7 @@ func (mc *MaterialController) UpdateMaterial(ctx *gin.Context) {
 
 	updatedMaterial, err := mc.MaterialUsecase.UpdateMaterial(ctx, material)
 	if err != nil {
+
 		ctx.JSON(500, gin.H{"error": "Failed to update material"})
 		return
 	}
@@ -251,6 +307,16 @@ func (mc *MaterialController) DeleteMaterial(ctx *gin.Context) {
 		return
 	}
 
+	material, err := mc.MaterialUsecase.GetMaterialByID(ctx, materialIDPrimitive)
+	if err != nil {
+		ctx.JSON(404, gin.H{"error": "Material not found"})
+		return
+	}
+	if material.UploadedBy != userIDPrimitive {
+		ctx.JSON(403, gin.H{"error": "You are not authorized to delete this material"})
+		return
+	}
+
 	err = mc.MaterialUsecase.DeleteMaterial(ctx, userIDPrimitive, materialIDPrimitive)
 
 	if err != nil {
@@ -258,18 +324,21 @@ func (mc *MaterialController) DeleteMaterial(ctx *gin.Context) {
 		return
 	}
 
+	mc.StorageUsecase.DeleteFile([]string{material.FileURL}) // Clean up the file from storage
+
 	ctx.JSON(204, gin.H{
 		"message": "Material deleted successfully",
 	})
 }
 
 func (mc *MaterialController) GetMaterialsInTree(ctx *gin.Context) {
+	fmt.Println(ctx.Request.URL)
 	departmentID := ctx.Query("department_id")
 	if departmentID == "" {
 		ctx.JSON(400, gin.H{"error": "Department ID is required"})
 		return
 	}
-
+	fmt.Println("Department ID:", departmentID)
 	departmentIDPrimitive, err := primitive.ObjectIDFromHex(departmentID)
 	if err != nil {
 		ctx.JSON(400, gin.H{"error": "Invalid Department ID format"})
@@ -292,6 +361,7 @@ func (mc *MaterialController) GetMaterialsInTree(ctx *gin.Context) {
 		ctx.JSON(400, gin.H{"error": "Semester is required"})
 		return
 	}
+	fmt.Println("Semester:", semesterStr, "Year:", year, "Department ID:", departmentID)
 	semester, err := strconv.Atoi(semesterStr)
 	if err != nil || (semester != 1 && semester != 2) {
 		ctx.JSON(400, gin.H{"error": "Invalid semester"})
